refactor: move calc functions into new highlight strategy file
