+++ src/chart_types/partition_chart/state/selectors/get_highlighted_shapes.ts
@@ -18,88 +18,22 @@
  */
 
 import createCachedSelector from 're-reselect';
-import { $Values } from 'utility-types';
 
-import { LegendPath } from '../../../../state/actions/legend';
 import { GlobalChartState } from '../../../../state/chart_state';
 import { getChartIdSelector } from '../../../../state/selectors/get_chart_id';
 import { getSettingsSpecSelector } from '../../../../state/selectors/get_settings_specs';
-import { DataName, QuadViewModel } from '../../layout/types/viewmodel_types';
+import { QuadViewModel } from '../../layout/types/viewmodel_types';
+import { highlightedGeoms } from '../../layout/utils/highlighted_geoms';
 import { partitionGeometries } from './geometries';
 
 const getHighlightedLegendItemPath = (state: GlobalChartState) => state.interactions.highlightedLegendPath;
 
-type LegendStrategyFn = (legendPath: LegendPath) => (partialShape: { path: LegendPath; dataName: DataName }) => boolean;
-
-const legendStrategies: Record<LegendStrategy, LegendStrategyFn> = {
-  node: (legendPath) => ({ path }) =>
-    // highlight exact match in the path only
-    legendPath.length === path.length &&
-    legendPath.every(({ index, value }, i) => index === path[i]?.index && value === path[i]?.value),
-
-  path: (legendPath) => ({ path }) =>
-    // highlight members of the exact path; ie. exact match in the path, plus all its ancestors
-    path.every(({ index, value }, i) => index === legendPath[i]?.index && value === legendPath[i]?.value),
-
-  keyInLayer: (legendPath) => ({ path, dataName }) =>
-    // highlight all identically named items which are within the same depth (ring) as the hovered legend depth
-    legendPath.length === path.length && dataName === legendPath[legendPath.length - 1].value,
-
-  key: (legendPath) => ({ dataName }) =>
-    // highlight all identically named items, no matter where they are
-    dataName === legendPath[legendPath.length - 1].value,
-
-  nodeWithDescendants: (legendPath) => ({ path }) =>
-    // highlight exact match in the path, and everything that is its descendant in that branch
-    legendPath.every(({ index, value }, i) => index === path[i]?.index && value === path[i]?.value),
-
-  pathWithDescendants: (legendPath) => ({ path }) =>
-    // highlight exact match in the path, and everything that is its ancestor, or its descendant in that branch
-    legendPath
-      .slice(0, path.length)
-      .every(({ index, value }, i) => index === path[i]?.index && value === path[i]?.value),
-};
-
-/** @public */
-export const LegendStrategy = Object.freeze({
-  /**
-   * Highlight the specific node(s) that the legend item stands for.
-   */
-  Node: 'node' as const,
-  /**
-   * Highlight members of the exact path; ie. like `Node`, plus all its ancestors
-   */
-  Path: 'path' as const,
-  /**
-   * Highlight all identically named (labelled) items within the tree layer (depth or ring) of the specific node(s) that the legend item stands for
-   */
-  KeyInLayer: 'keyInLayer' as const,
-  /**
-   * Highlight all identically named (labelled) items, no matter where they are
-   */
-  Key: 'key' as const,
-  /**
-   * Highlight the specific node(s) that the legend item stands for, plus all descendants
-   */
-  NodeWithDescendants: 'nodeWithDescendants' as const,
-  /**
-   * Highlight the specific node(s) that the legend item stands for, plus all ancestors and descendants
-   */
-  PathWithDescendants: 'pathWithDescendants' as const,
-});
-
-/** @public */
-export type LegendStrategy = $Values<typeof LegendStrategy>;
-const defaultStrategy: LegendStrategy = LegendStrategy.Key;
-
 /** @internal */
 export const legendHoverHighlightNodes = createCachedSelector(
   [getSettingsSpecSelector, getHighlightedLegendItemPath, partitionGeometries],
-  (specs, highlightedLegendItemPath, geoms): QuadViewModel[] => {
-    if (highlightedLegendItemPath.length === 0) {
-      return [];
-    }
-    const pickedLogic: LegendStrategy = specs.legendStrategy ?? defaultStrategy;
-    return geoms.quadViewModel.filter(legendStrategies[pickedLogic](highlightedLegendItemPath));
+  ({ legendStrategy }, highlightedLegendItemPath, { quadViewModel }): QuadViewModel[] => {
+    return highlightedLegendItemPath.length > 0
+      ? highlightedGeoms(legendStrategy, quadViewModel, highlightedLegendItemPath)
+      : [];
   },
 )(getChartIdSelector);
